[AWIBOF-4100] Add unit for volume list command
Add unit (B, KB, MB, GB ...) when executing volume list command.


Former-commit-id: 4d24054b9b8346d045c9d27f1e35442fd5ab286a

+++ tool/cli/cmd/messages/responses.go
@@ -160,8 +160,8 @@ type ListVolumeResData struct {
 type Volume struct {
 	VOLUMENAME string `json:"name"`
 	VOLUMEID   int    `json:"id"`
-	TOTAL      int    `json:"total"`
-	REMAIN     int    `json:"remain"`
+	TOTAL      uint64 `json:"total"`
+	REMAIN     uint64 `json:"remain"`
 	STATUS     string `json:"status"`
 	MAXIOPS    int    `json:"maxiops"`
 	MAXBW      int    `json:"maxbw"`
